Update direct binding from sol to go using abigen
The new abigen version does not have the --solc file to directly bind a Solidity contract to a Go package. This was caused because the solc compiler in ```go-ethereum``` needed to be continuously synchronized with the original solc compiler in order to offer the full features and security of latest versions. The team decided to remove the direct biding functionality from Solidity to Go but the documentation still has it (therefore confusing newcomers with outdated instructions).